chore: rename to `ais-search-state`
