+++ src/__tests__/index.js
@@ -26,8 +26,8 @@ test('Should register all components when installed', () => {
     'ais-range-input',
     'ais-rating-menu',
     'ais-refinement-list',
-    'ais-results',
     'ais-search-box',
+    'ais-search-state',
     'ais-snippet',
     'ais-sort-by',
     'ais-stats',
